+++ src/main/java/dk/fantastiskefroe/spir/ingest/entity/OrderLine.java
@@ -13,5 +13,5 @@ public record OrderLine(
         @NonNull
         int quantity,
         @Nullable
-        int price) {
+        Double price) {
 }
